test(compose): the deep read's acceptance rules become their own file (#875)
`main` red #4. deepread_integration_test.go reached 1012 lines and the strict
craft bar stops at 1000, so the craftsmanship job blocks every PR — it does not
run on main's own pushes, which is why main looks green on it and nothing else
can merge.

Split where the file already had a seam rather than at the line count. What the
worker DOES — crawl, extract, stage one proposal, record an honest outcome,
survive a failure midway — is one subject; what the record BECOMES once a human
answers that proposal is another. The three acceptance rules move: offerings
dedupe on their value key and never overwrite a human's value, an accepted
employee_range fills the size band only where the mapping is unambiguous, and a
rejection lands nothing.

772 and 268 lines, 15 tests and 3, all eighteen still present and passing
against a live database. Imports on both sides are derived from what each half
actually uses.
